included ratio prompt/frag from 8 TeV pythia in theory compilation
I included the spectra of prompt photons from Pythia8, fragm. photons (only from JJ processes), their ratio and respective fits. The ratio is to be used in the merged analysis for the purity correction of photons. The new .root files also contain several other spectra, but they are not included in the theory compilation at the moment.